finish migration of grab from robot -> kinbody, fixed bug in CheckEndEffectorCollision
