+++ include/openrave/robot.h
@@ -892,44 +892,6 @@ private:
      */
     virtual void SimulationStep(dReal fElapsedTime);
 
-    /** \brief Check if body is self colliding with its links or its grabbed bodies.
-
-        Links that are joined together are ignored.
-        Collisions between grabbed bodies are also considered as self-collisions for this body.
-        \param report [optional] collision report
-     */
-    virtual bool CheckSelfCollision(CollisionReportPtr report = CollisionReportPtr(), CollisionCheckerBasePtr collisionchecker=CollisionCheckerBasePtr()) const;
-
-    /** \brief checks collision of a robot link with the surrounding environment using a new transform. Attached/Grabbed bodies to this link are also checked for collision.
-
-        \param[in] ilinkindex the index of the link to check
-        \param[in] tlinktrans The transform of the link to check
-        \param[out] report [optional] collision report
-     */
-    virtual bool CheckLinkCollision(int ilinkindex, const Transform& tlinktrans, CollisionReportPtr report = CollisionReportPtr());
-
-    /** \brief checks collision of a robot link with the surrounding environment using the current link's transform. Attached/Grabbed bodies to this link are also checked for collision.
-
-        \param[in] ilinkindex the index of the link to check
-        \param[out] report [optional] collision report
-     */
-    virtual bool CheckLinkCollision(int ilinkindex, CollisionReportPtr report = CollisionReportPtr());
-
-    /** \brief checks self-collision of a robot link with the other robot links. Attached/Grabbed bodies to this link are also checked for self-collision.
-
-        \param[in] ilinkindex the index of the link to check
-        \param[out] report [optional] collision report
-     */
-    virtual bool CheckLinkSelfCollision(int ilinkindex, CollisionReportPtr report = CollisionReportPtr());
-    
-    /** \brief checks self-collision of a robot link with the other robot links. Attached/Grabbed bodies to this link are also checked for self-collision.
-
-        \param[in] ilinkindex the index of the link to check
-        \param[in] tlinktrans The transform of the link to check
-        \param[out] report [optional] collision report
-     */
-    virtual bool CheckLinkSelfCollision(int ilinkindex, const Transform& tlinktrans, CollisionReportPtr report = CollisionReportPtr());
-
     /// does not clone the grabbed bodies since it requires pointers from other bodies (that might not be initialized yet)
     virtual void Clone(InterfaceBaseConstPtr preference, int cloningoptions);
