+++ src/libopenrave/kinbody.cpp
@@ -60,20 +60,6 @@ protected:
 
 typedef boost::shared_ptr<ChangeCallbackData> ChangeCallbackDataPtr;
 
-
-class CallOnDestruction
-{
-public:
-    CallOnDestruction(const boost::function<void()>& fn) : _fn(fn) {
-    }
-    ~CallOnDestruction() {
-        _fn();
-    }
-private:
-    boost::function<void()> _fn;
-};
-
-
 ElectricMotorActuatorInfo::ElectricMotorActuatorInfo()
 {
     gear_ratio = 0;
@@ -248,6 +234,7 @@ KinBody::~KinBody()
 
 void KinBody::Destroy()
 {
+    ReleaseAllGrabbed();
     if( _listAttachedBodies.size() > 0 ) {
         // could be in the environment destructor?
         stringstream ss; ss << GetName() << " still has attached bodies: ";
@@ -945,6 +932,7 @@ void KinBody::SetDOFTorqueLimits(const std::vector<dReal>& v)
 
 void KinBody::SimulationStep(dReal fElapsedTime)
 {
+    _UpdateGrabbedBodies();
 }
 
 void KinBody::SubtractDOFValues(std::vector<dReal>& q1, const std::vector<dReal>& q2, const std::vector<int>& dofindices) const
@@ -958,7 +946,7 @@ void KinBody::SubtractDOFValues(std::vector<dReal>& q1, const std::vector<dReal>
     }
 
     if( dofindices.size() == 0 ) {
-        OPENRAVE_ASSERT_OP(q1.size(), ==, GetDOF() );
+        OPENRAVE_ASSERT_OP((int)q1.size(), ==, GetDOF() );
         FOREACHC(itjoint,_vecjoints) {
             int dof = (*itjoint)->GetDOFIndex();
             for(int i = 0; i < (*itjoint)->GetDOF(); ++i) {
@@ -997,6 +985,8 @@ void KinBody::SetTransform(const Transform& trans)
     FOREACH(itlink, _veclinks) {
         (*itlink)->SetTransform(tapply * (*itlink)->GetTransform());
     }
+    _UpdateGrabbedBodies();
+    _PostprocessChangedParameters(Prop_LinkTransforms);
 }
 
 Transform KinBody::GetTransform() const
@@ -1015,7 +1005,10 @@ bool KinBody::SetVelocity(const Vector& linearvel, const Vector& angularvel)
             velocities[i].first = linearvel + angularvel.cross(_veclinks[i]->GetTransform().trans-vlinktrans);
             velocities[i].second = angularvel;
         }
-        return GetEnv()->GetPhysicsEngine()->SetLinkVelocities(shared_kinbody(),velocities);
+
+        bool bSuccess = GetEnv()->GetPhysicsEngine()->SetLinkVelocities(shared_kinbody(),velocities);
+        _UpdateGrabbedBodies();
+        return bSuccess;
     }
     return false;
 }
@@ -1445,6 +1438,7 @@ void KinBody::SetLinkTransformations(const std::vector<Transform>& vbodies)
             (*itjoint)->_doflastsetvalues[i] = (*itjoint)->GetValue(i);
         }
     }
+    _UpdateGrabbedBodies();
     _PostprocessChangedParameters(Prop_LinkTransforms);
 }
 
@@ -1461,12 +1455,14 @@ void KinBody::SetLinkTransformations(const std::vector<Transform>& transforms, c
             (*itjoint)->_doflastsetvalues[i] = doflastsetvalues.at((*itjoint)->GetDOFIndex()+i);
         }
     }
+    _UpdateGrabbedBodies();
     _PostprocessChangedParameters(Prop_LinkTransforms);
 }
 
 void KinBody::SetLinkVelocities(const std::vector<std::pair<Vector,Vector> >& velocities)
 {
     GetEnv()->GetPhysicsEngine()->SetLinkVelocities(shared_kinbody(),velocities);
+    _UpdateGrabbedBodies();
 }
 
 void KinBody::SetLinkEnableStates(const std::vector<uint8_t>& enablestates)
@@ -1829,6 +1825,7 @@ void KinBody::SetDOFValues(const std::vector<dReal>& vJointValues, uint32_t chec
         vlinkscomputed[pjoint->GetHierarchyChildLink()->GetIndex()] = 1;
     }
 
+    _UpdateGrabbedBodies();
     _PostprocessChangedParameters(Prop_LinkTransforms);
 }
 
@@ -3331,46 +3328,6 @@ CollisionCheckerBasePtr KinBody::GetSelfCollisionChecker() const
     return _selfcollisionchecker;
 }
 
-bool KinBody::CheckSelfCollision(CollisionReportPtr report, CollisionCheckerBasePtr collisionchecker) const
-{
-    if( !collisionchecker ) {
-        collisionchecker = _selfcollisionchecker;
-        if( !collisionchecker ) {
-            collisionchecker = GetEnv()->GetCollisionChecker();
-            if( !collisionchecker ) {
-                // no checker set
-                return false;
-            }
-        }
-        else {
-            // have to set the same options as GetEnv()->GetCollisionChecker() since stuff like CO_ActiveDOFs is only set on the global checker
-            collisionchecker->SetCollisionOptions(GetEnv()->GetCollisionChecker()->GetCollisionOptions());
-        }
-    }
-
-    if( collisionchecker->CheckStandaloneSelfCollision(shared_kinbody_const(), report) ) {
-        if( !!report ) {
-            if( IS_DEBUGLEVEL(Level_Verbose) ) {
-                std::vector<OpenRAVE::dReal> v;
-                GetDOFValues(v);
-                stringstream ss; ss << std::setprecision(std::numeric_limits<OpenRAVE::dReal>::digits10+1);
-                ss << "self collision report=" << report->__str__() << " ";
-                for(size_t i = 0; i < v.size(); ++i ) {
-                    if( i > 0 ) {
-                        ss << "," << v[i];
-                    }
-                    else {
-                        ss << "colvalues=[" << v[i];
-                    }
-                }
-                ss << "]";
-                RAVELOG_VERBOSE(ss.str());
-            }
-        }
-        return true;
-    }
-    return false;
-}
 
 void KinBody::_ComputeInternalInformation()
 {
@@ -4426,29 +4383,6 @@ void KinBody::SetNonCollidingConfiguration()
     GetLinkTransformations(_vInitialLinkTransformations, vdoflastsetvalues);
 }
 
-void KinBody::_UpdateGrabbedBodies()
-{
-    vector<UserDataPtr>::iterator itgrabbed = _vGrabbedBodies.begin();
-    while(itgrabbed != _vGrabbedBodies.end() ) {
-        GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
-        KinBodyPtr pbody = pgrabbed->_pgrabbedbody.lock();
-        if( !!pbody ) {
-            Transform t = pgrabbed->_plinkrobot->GetTransform();
-            pbody->SetTransform(t * pgrabbed->_troot);
-            // set the correct velocity
-            std::pair<Vector, Vector> velocity = pgrabbed->_plinkrobot->GetVelocity();
-            velocity.first += velocity.second.cross(t.rotate(pgrabbed->_troot.trans));
-            pbody->SetVelocity(velocity.first, velocity.second);
-            ++itgrabbed;
-        }
-        else {
-            RAVELOG_DEBUG(str(boost::format("erasing invaliding grabbed body from %s")%GetName()));
-            itgrabbed = _vGrabbedBodies.erase(itgrabbed);
-        }
-    }
-}
-
-    
 void KinBody::_ResetInternalCollisionCache()
 {
     _nNonAdjacentLinkCache = 0x80000000;
@@ -4685,6 +4619,28 @@ void KinBody::Clone(InterfaceBaseConstPtr preference, int cloningoptions)
 
     // cache
     _ResetInternalCollisionCache();
+
+    // clone the grabbed bodies, note that this can fail if the new cloned environment hasn't added the bodies yet (check out Environment::Clone)
+    _vGrabbedBodies.resize(0);
+    FOREACHC(itgrabbedref, r->_vGrabbedBodies) {
+        GrabbedConstPtr pgrabbedref = boost::dynamic_pointer_cast<Grabbed const>(*itgrabbedref);
+
+        KinBodyPtr pbodyref(pgrabbedref->_pgrabbedbody);
+        KinBodyPtr pgrabbedbody;
+        if( !!pbodyref ) {
+            pgrabbedbody = GetEnv()->GetBodyFromEnvironmentId(pbodyref->GetEnvironmentId());
+            BOOST_ASSERT(pgrabbedbody->GetName() == pbodyref->GetName());
+
+            GrabbedPtr pgrabbed(new Grabbed(pgrabbedbody,_veclinks.at(KinBody::LinkPtr(pgrabbedref->_plinkrobot)->GetIndex())));
+            pgrabbed->_troot = pgrabbedref->_troot;
+            pgrabbed->_listNonCollidingLinks.clear();
+            FOREACHC(itlinkref, pgrabbedref->_listNonCollidingLinks) {
+                pgrabbed->_listNonCollidingLinks.push_back(_veclinks.at((*itlinkref)->GetIndex()));
+            }
+            _vGrabbedBodies.push_back(pgrabbed);
+        }
+    }
+
     _nUpdateStampId++; // update the stamp instead of copying
 }
 
@@ -4705,6 +4661,60 @@ void KinBody::_PostprocessChangedParameters(uint32_t parameters)
     // do not change hash if geometry changed!
     if( !!(parameters & (Prop_LinkDynamics|Prop_LinkGeometry|Prop_JointMimic)) ) {
         __hashkinematics.resize(0);
+    }
+
+    if( (parameters&Prop_LinkEnable) == Prop_LinkEnable ) {
+        // check if any regrabbed bodies have the link in _listNonCollidingLinks and the link is enabled, or are missing the link in _listNonCollidingLinks and the link is disabled
+        std::map<GrabbedPtr, list<KinBody::LinkConstPtr> > mapcheckcollisions;
+        FOREACH(itlink,_veclinks) {
+            if( (*itlink)->IsEnabled() ) {
+                FOREACH(itgrabbed,_vGrabbedBodies) {
+                    GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
+                    if( find(pgrabbed->GetRigidlyAttachedLinks().begin(),pgrabbed->GetRigidlyAttachedLinks().end(), *itlink) == pgrabbed->GetRigidlyAttachedLinks().end() ) {
+                        std::list<KinBody::LinkConstPtr>::iterator itnoncolliding = find(pgrabbed->_listNonCollidingLinks.begin(),pgrabbed->_listNonCollidingLinks.end(),*itlink);
+                        if( itnoncolliding != pgrabbed->_listNonCollidingLinks.end() ) {
+                            if( pgrabbed->WasLinkNonColliding(*itlink) == 0 ) {
+                                pgrabbed->_listNonCollidingLinks.erase(itnoncolliding);
+                            }
+                            mapcheckcollisions[pgrabbed].push_back(*itlink);
+                        }
+                        else {
+                            // try to restore
+                            if( pgrabbed->WasLinkNonColliding(*itlink) == 1 ) {
+                                pgrabbed->_listNonCollidingLinks.push_back(*itlink);
+                            }
+                        }
+                    }
+                }
+            }
+            else {
+                // add since it is disabled?
+                FOREACH(itgrabbed,_vGrabbedBodies) {
+                    GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
+                    if( find(pgrabbed->GetRigidlyAttachedLinks().begin(),pgrabbed->GetRigidlyAttachedLinks().end(), *itlink) == pgrabbed->GetRigidlyAttachedLinks().end() ) {
+                        if( find(pgrabbed->_listNonCollidingLinks.begin(),pgrabbed->_listNonCollidingLinks.end(),*itlink) == pgrabbed->_listNonCollidingLinks.end() ) {
+                            if( pgrabbed->WasLinkNonColliding(*itlink) != 0 ) {
+                                pgrabbed->_listNonCollidingLinks.push_back(*itlink);
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+//        if( mapcheckcollisions.size() > 0 ) {
+//            CollisionOptionsStateSaver colsaver(GetEnv()->GetCollisionChecker(),0); // have to reset the collision options
+//            FOREACH(itgrabbed, mapcheckcollisions) {
+//                KinBodyPtr pgrabbedbody(itgrabbed->first->_pgrabbedbody);
+//                _RemoveAttachedBody(pgrabbedbody);
+//                CallOnDestruction destructionhook(boost::bind(&RobotBase::_AttachBody,this,pgrabbedbody));
+//                FOREACH(itlink, itgrabbed->second) {
+//                    if( pchecker->CheckCollision(*itlink, KinBodyConstPtr(pgrabbedbody)) ) {
+//                        itgrabbed->first->_listNonCollidingLinks.remove(*itlink);
+//                    }
+//                }
+//            }
+//        }
     }
 
     std::list<UserDataWeakPtr> listRegisteredCallbacks;
@@ -4820,291 +4830,6 @@ ConfigurationSpecification KinBody::GetConfigurationSpecificationIndices(const s
         spec._vgroups[0].interpolation=interpolation;
     }
     return spec;
-}
-
-bool KinBody::Grab(KinBodyPtr pbody, LinkPtr plink)
-{
-    OPENRAVE_ASSERT_FORMAT(!!pbody, "body %s invalid body to grab",GetName(),ORE_InvalidArguments);
-    OPENRAVE_ASSERT_FORMAT(!!plink && plink->GetParent() == shared_kinbody(), "body %s grabbing link needs to be part of body",GetName(),ORE_InvalidArguments);
-    OPENRAVE_ASSERT_FORMAT(pbody != shared_kinbody(),"body %s cannot grab itself",GetName(), ORE_InvalidArguments);
-
-    //uint64_t starttime0 = utils::GetMicroTime();
-
-    // if grabbing, check if the transforms are different. If they are, then update the transform
-    GrabbedPtr pPreviousGrabbed;
-    FOREACHC(itgrabbed, _vGrabbedBodies) {
-        GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
-        if( KinBodyConstPtr(pgrabbed->_pgrabbedbody) == pbody ) {
-            pPreviousGrabbed = pgrabbed;
-            break;
-        }
-    }
-
-    // double check since collision checkers might not support this case
-    if( pbody->HasAttached() ) {
-        RAVELOG_WARN_FORMAT("body %s trying to grab body %s with %d attached bodies", GetName()%pbody->GetName()%pbody->HasAttached());
-    }
-
-    Transform t = plink->GetTransform();
-    Transform tbody = pbody->GetTransform();
-    // new body velocity is measured from body link
-    std::pair<Vector, Vector> velocity = plink->GetVelocity();
-    velocity.first += velocity.second.cross(tbody.trans - t.trans);
-    if( !!pPreviousGrabbed ) {
-        dReal disterror = TransformDistance2(t*pPreviousGrabbed->_troot, tbody);
-        if( pPreviousGrabbed->_plinkrobot == plink && disterror <= g_fEpsilonLinear ) {
-            // links and transforms are the same, so no worries
-            return true;
-        }
-        RAVELOG_VERBOSE_FORMAT("Body %s: body %s already grabbed, but transforms differ by %f \n", GetName()%pbody->GetName()%disterror);
-        _RemoveAttachedBody(pbody);
-        CallOnDestruction destructigonhook(boost::bind(&RobotBase::_AttachBody,this,pbody));
-        pPreviousGrabbed->_plinkrobot = plink;
-        pPreviousGrabbed->_troot = t.inverse() * tbody;
-        pPreviousGrabbed->_ProcessCollidingLinks(pPreviousGrabbed->_setRobotLinksToIgnore);
-        pbody->SetVelocity(velocity.first, velocity.second);
-        return true;
-    }
-
-    GrabbedPtr pgrabbed(new Grabbed(pbody,plink));
-    pgrabbed->_troot = t.inverse() * tbody;
-    //uint64_t starttime1 = utils::GetMicroTime();
-    // always ignore links that are statically attached to plink (ie assume they are always colliding with the body)
-
-    std::vector<boost::shared_ptr<Link> > vattachedlinks;
-    plink->GetRigidlyAttachedLinks(vattachedlinks);
-    std::set<int> setBodyLinksToIgnore;
-    FOREACHC(itlink, vattachedlinks) {
-        setBodyLinksToIgnore.insert((*itlink)->GetIndex());
-    }
-    if( !!_selfcollisionchecker && _selfcollisionchecker != GetEnv()->GetCollisionChecker() ) {
-        // collision checking will not be automatically updated with environment calls, so need to do this manually
-        _selfcollisionchecker->InitKinBody(pbody);
-    }
-    pgrabbed->_ProcessCollidingLinks(setBodyLinksToIgnore);
-    pbody->SetVelocity(velocity.first, velocity.second);
-    _vGrabbedBodies.push_back(pgrabbed);
-    //uint64_t starttime2 = utils::GetMicroTime();
-    try {
-        // if an exception happens in _AttachBody, have to remove from _vGrabbedBodies
-        _AttachBody(pbody);
-    }
-    catch(...) {
-        BOOST_ASSERT(_vGrabbedBodies.back()==pgrabbed);
-        // do not call _selfcollisionchecker->RemoveKinBody since the same object might be re-attached later on and we should preserve the structures.
-        _vGrabbedBodies.pop_back();
-        throw;
-    }
-    //uint64_t starttime3 = utils::GetMicroTime();
-    _PostprocessChangedParameters(Prop_RobotGrabbed);
-    //RAVELOG_DEBUG_FORMAT("env=%d, post process elapsed (%d) %fs, %fs, %fs, %fs", GetEnv()->GetId()%vattachedlinks.size()%(1e-6*(starttime1-starttime0))%(1e-6*(starttime2-starttime0))%(1e-6*(starttime3-starttime0))%(1e-6*(utils::GetMicroTime()-starttime0)));
-    return true;
-}
-
-bool KinBody::Grab(KinBodyPtr pbody, LinkPtr pBodyLinkToGrabWith, const std::set<int>& setBodyLinksToIgnore)
-{
-    OPENRAVE_ASSERT_FORMAT(!!pbody && !!pBodyLinkToGrabWith && pBodyLinkToGrabWith->GetParent() == shared_kinbody(), "body %s invalid grab arguments",GetName(), ORE_InvalidArguments);
-    OPENRAVE_ASSERT_FORMAT(pbody != shared_kinbody(), "body %s cannot grab itself",pbody->GetName(), ORE_InvalidArguments);
-    if( IsGrabbing(pbody) ) {
-        if( setBodyLinksToIgnore.size() > 0 ) {
-            // update the current grabbed info with setBodyLinksToIgnore
-            FOREACHC(itgrabbed, _vGrabbedBodies) {
-                GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
-                if( KinBodyConstPtr(pgrabbed->_pgrabbedbody) == pbody ) {
-                    pgrabbed->AddMoreIgnoreLinks(setBodyLinksToIgnore);
-                    break;
-                }
-            }
-        }
-        RAVELOG_VERBOSE(str(boost::format("Body %s: body %s already grabbed\n")%GetName()%pbody->GetName()));
-        return true;
-    }
-
-    GrabbedPtr pgrabbed(new Grabbed(pbody,pBodyLinkToGrabWith));
-    Transform t = pBodyLinkToGrabWith->GetTransform();
-    Transform tbody = pbody->GetTransform();
-    pgrabbed->_troot = t.inverse() * tbody;
-
-    if( !!_selfcollisionchecker && _selfcollisionchecker != GetEnv()->GetCollisionChecker() ) {
-        // collision checking will not be automatically updated with environment calls, so need to do this manually
-        _selfcollisionchecker->InitKinBody(pbody);
-    }
-    pgrabbed->_ProcessCollidingLinks(setBodyLinksToIgnore);
-
-    // set velocity
-    std::pair<Vector, Vector> velocity = pBodyLinkToGrabWith->GetVelocity();
-    velocity.first += velocity.second.cross(tbody.trans - t.trans);
-    pbody->SetVelocity(velocity.first, velocity.second);
-    _vGrabbedBodies.push_back(pgrabbed);
-    try {
-        // if an exception happens in _AttachBody, have to remove from _vGrabbedBodies
-        _AttachBody(pbody);
-    }
-    catch(...) {
-        BOOST_ASSERT(_vGrabbedBodies.back()==pgrabbed);
-        // do not call _selfcollisionchecker->RemoveKinBody since the same object might be re-attached later on and we should preserve the structures.
-        _vGrabbedBodies.pop_back();
-        throw;
-    }
-    _PostprocessChangedParameters(Prop_RobotGrabbed);
-    return true;
-}
-
-void KinBody::Release(KinBodyPtr pbody)
-{
-    FOREACH(itgrabbed, _vGrabbedBodies) {
-        GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
-        if( KinBodyPtr(pgrabbed->_pgrabbedbody) == pbody ) {
-            _vGrabbedBodies.erase(itgrabbed);
-            _RemoveAttachedBody(pbody);
-            _PostprocessChangedParameters(Prop_RobotGrabbed);
-            return;
-        }
-    }
-
-    RAVELOG_DEBUG(str(boost::format("Body %s: body %s not grabbed\n")%GetName()%pbody->GetName()));
-}
-
-void KinBody::ReleaseAllGrabbed()
-{
-    if( _vGrabbedBodies.size() > 0 ) {
-        FOREACH(itgrabbed, _vGrabbedBodies) {
-            GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
-            KinBodyPtr pbody = pgrabbed->_pgrabbedbody.lock();
-            if( !!pbody ) {
-                _RemoveAttachedBody(pbody);
-            }
-        }
-        _vGrabbedBodies.clear();
-        _PostprocessChangedParameters(Prop_RobotGrabbed);
-    }
-}
-
-void KinBody::RegrabAll()
-{
-    CollisionCheckerBasePtr collisionchecker = !!_selfcollisionchecker ? _selfcollisionchecker : GetEnv()->GetCollisionChecker();
-    CollisionOptionsStateSaver colsaver(collisionchecker,0); // have to reset the collision options
-    std::vector<LinkPtr > vattachedlinks;
-    FOREACH(itgrabbed, _vGrabbedBodies) {
-        GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
-        KinBodyPtr pbody(pgrabbed->_pgrabbedbody);
-        if( !!pbody ) {
-            _RemoveAttachedBody(pbody);
-            CallOnDestruction destructionhook(boost::bind(&RobotBase::_AttachBody,this,pbody));
-            pgrabbed->_ProcessCollidingLinks(pgrabbed->_setRobotLinksToIgnore);
-        }
-    }
-}
-
-void KinBody::_Regrab(UserDataPtr _pgrabbed)
-{
-    GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(_pgrabbed);
-    KinBodyPtr pgrabbedbody = pgrabbed->_pgrabbedbody.lock();
-    if( !!pgrabbedbody ) {
-        // have to re-grab the body, which means temporarily resetting the collision checker and attachment
-        CollisionCheckerBasePtr collisionchecker = !!_selfcollisionchecker ? _selfcollisionchecker : GetEnv()->GetCollisionChecker();
-        CollisionOptionsStateSaver colsaver(collisionchecker,0); // have to reset the collision options
-        _RemoveAttachedBody(pgrabbedbody);
-        CallOnDestruction destructionhook(boost::bind(&RobotBase::_AttachBody,this,pgrabbedbody));
-        pgrabbed->_ProcessCollidingLinks(pgrabbed->_setRobotLinksToIgnore);
-    }
-}
-
-KinBody::LinkPtr KinBody::IsGrabbing(KinBodyConstPtr pbody) const
-{
-    FOREACHC(itgrabbed, _vGrabbedBodies) {
-        GrabbedPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed>(*itgrabbed);
-        if( KinBodyConstPtr(pgrabbed->_pgrabbedbody) == pbody ) {
-            return pgrabbed->_plinkrobot;
-        }
-    }
-    return LinkPtr();
-}
-
-void KinBody::GetGrabbed(std::vector<KinBodyPtr>& vbodies) const
-{
-    vbodies.resize(0);
-    FOREACHC(itgrabbed, _vGrabbedBodies) {
-        GrabbedConstPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed const>(*itgrabbed);
-        KinBodyPtr pbody = pgrabbed->_pgrabbedbody.lock();
-        if( !!pbody && pbody->GetEnvironmentId() ) {
-            vbodies.push_back(pbody);
-        }
-    }
-}
-
-void KinBody::GetGrabbedInfo(std::vector<KinBody::GrabbedInfoPtr>& vgrabbedinfo) const
-{
-    vgrabbedinfo.resize(_vGrabbedBodies.size());
-    for(size_t i = 0; i < vgrabbedinfo.size(); ++i) {
-        GrabbedConstPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed const>(_vGrabbedBodies[i]);
-        vgrabbedinfo[i].reset(new GrabbedInfo());
-        vgrabbedinfo[i]->_grabbedname = pgrabbed->_pgrabbedbody.lock()->GetName();
-        vgrabbedinfo[i]->_robotlinkname = pgrabbed->_plinkrobot->GetName();
-        vgrabbedinfo[i]->_trelative = pgrabbed->_troot;
-        vgrabbedinfo[i]->_setRobotLinksToIgnore = pgrabbed->_setRobotLinksToIgnore;
-        FOREACHC(itlink, _veclinks) {
-            if( find(pgrabbed->_listNonCollidingLinks.begin(), pgrabbed->_listNonCollidingLinks.end(), *itlink) == pgrabbed->_listNonCollidingLinks.end() ) {
-                vgrabbedinfo[i]->_setRobotLinksToIgnore.insert((*itlink)->GetIndex());
-            }
-        }
-    }
-}
-
-void KinBody::ResetGrabbed(const std::vector<KinBody::GrabbedInfoConstPtr>& vgrabbedinfo)
-{
-    ReleaseAllGrabbed();
-    if( vgrabbedinfo.size() > 0 ) {
-        CollisionCheckerBasePtr collisionchecker = !!_selfcollisionchecker ? _selfcollisionchecker : GetEnv()->GetCollisionChecker();
-        CollisionOptionsStateSaver colsaver(collisionchecker,0); // have to reset the collision options
-        FOREACHC(itgrabbedinfo, vgrabbedinfo) {
-            GrabbedInfoConstPtr pgrabbedinfo = *itgrabbedinfo;
-            KinBodyPtr pbody = GetEnv()->GetKinBody(pgrabbedinfo->_grabbedname);
-            KinBody::LinkPtr pBodyLinkToGrabWith = GetLink(pgrabbedinfo->_robotlinkname);
-            OPENRAVE_ASSERT_FORMAT(!!pbody && !!pBodyLinkToGrabWith, "body %s invalid grab arguments",GetName(), ORE_InvalidArguments);
-            OPENRAVE_ASSERT_FORMAT(pbody != shared_kinbody(), "body %s cannot grab itself",pbody->GetName(), ORE_InvalidArguments);
-            if( IsGrabbing(pbody) ) {
-                RAVELOG_VERBOSE(str(boost::format("Body %s: body %s already grabbed\n")%GetName()%pbody->GetName()));
-                continue;
-            }
-
-            GrabbedPtr pgrabbed(new Grabbed(pbody,pBodyLinkToGrabWith));
-            pgrabbed->_troot = pgrabbedinfo->_trelative;
-            if( !!_selfcollisionchecker && _selfcollisionchecker != GetEnv()->GetCollisionChecker() ) {
-                // collision checking will not be automatically updated with environment calls, so need to do this manually
-                _selfcollisionchecker->InitKinBody(pbody);
-            }
-            pgrabbed->_ProcessCollidingLinks(pgrabbedinfo->_setRobotLinksToIgnore);
-            Transform tlink = pBodyLinkToGrabWith->GetTransform();
-            Transform tbody = tlink * pgrabbed->_troot;
-            pbody->SetTransform(tbody);
-            // set velocity
-            std::pair<Vector, Vector> velocity = pBodyLinkToGrabWith->GetVelocity();
-            velocity.first += velocity.second.cross(tbody.trans - tlink.trans);
-            pbody->SetVelocity(velocity.first, velocity.second);
-            _vGrabbedBodies.push_back(pgrabbed);
-            _AttachBody(pbody);
-        }
-        _PostprocessChangedParameters(Prop_RobotGrabbed);
-    }
-}
-
-void KinBody::GetIgnoredLinksOfGrabbed(KinBodyConstPtr body, std::list<KinBody::LinkConstPtr>& ignorelinks) const
-{
-    ignorelinks.clear();
-    FOREACHC(itgrabbed, _vGrabbedBodies) {
-        GrabbedConstPtr pgrabbed = boost::dynamic_pointer_cast<Grabbed const>(*itgrabbed);
-        KinBodyPtr grabbedbody = pgrabbed->_pgrabbedbody.lock();
-        if( grabbedbody == body ) {
-            FOREACHC(itbodylink, _veclinks) {
-                if( find(pgrabbed->_listNonCollidingLinks.begin(), pgrabbed->_listNonCollidingLinks.end(), *itbodylink) == pgrabbed->_listNonCollidingLinks.end() ) {
-                    ignorelinks.push_back(*itbodylink);
-                }
-            }
-            return;
-        }
-    }
-    RAVELOG_WARN(str(boost::format("body %s is not currently grabbed")%body->GetName()));
 }
 
 UserDataPtr KinBody::RegisterChangeCallback(uint32_t properties, const boost::function<void()>&callback) const
