+++ src/libopenrave/libopenrave.cpp
@@ -480,7 +480,7 @@ public:
         if( !!pOPENRAVE_DEFAULT_VIEWER && strlen(pOPENRAVE_DEFAULT_VIEWER) > 0 ) {
             _defaultviewertype = std::string(pOPENRAVE_DEFAULT_VIEWER);
         }
-        
+
         _UpdateDataDirs();
         return 0;
     }
@@ -491,7 +491,7 @@ public:
             // notify all plugins that about to destroy
             _pdatabase->OnRavePreDestroy();
         }
-        
+
         // don't use any log statements since global instance might be null
         // environments have to be destroyed carefully since their destructors can be called, which will attempt to unregister the environment
         std::map<int, EnvironmentBase*> mapenvironments;
@@ -842,7 +842,7 @@ protected:
         if( _defaultviewertype.size() > 0 ) {
             return _defaultviewertype;
         }
-        
+
         // get the first viewer that can be loadable, with preferenace to qtosg, qtcoin
         boost::shared_ptr<RaveDatabase> pdatabase = _pdatabase;
         if( !!pdatabase ) {
@@ -2067,8 +2067,7 @@ void TriMesh::serialize(std::ostream& o, int options) const
     }
 }
 
-
-void Grabbed::_ProcessCollidingLinks(const std::set<int>& setRobotLinksToIgnore)
+void Grabbed::ProcessCollidingLinks(const std::set<int>& setRobotLinksToIgnore)
 {
     _setRobotLinksToIgnore = setRobotLinksToIgnore;
     _listNonCollidingLinks.clear();
@@ -2143,6 +2142,112 @@ void Grabbed::_ProcessCollidingLinks(const std::set<int>& setRobotLinksToIgnore)
                 _listNonCollidingLinks.push_back(itnoncolliding->first);
             }
         }
+    }
+}
+
+void Grabbed::AddMoreIgnoreLinks(const std::set<int>& setRobotLinksToIgnore)
+{
+    KinBodyPtr pbody = RaveInterfaceCast<KinBody>(_plinkrobot->GetParent());
+    FOREACHC(itignoreindex, setRobotLinksToIgnore) {
+        _setRobotLinksToIgnore.insert(*itignoreindex);
+        KinBody::LinkPtr plink = pbody->GetLinks().at(*itignoreindex);
+        _mapLinkIsNonColliding[plink] = 0;
+        _listNonCollidingLinks.remove(plink);
+    }
+}
+
+/// return -1 for unknown, 0 for no, 1 for yes
+int Grabbed::WasLinkNonColliding(KinBody::LinkConstPtr plink) const
+{
+    std::map<KinBody::LinkConstPtr, int>::const_iterator it = _mapLinkIsNonColliding.find(plink);
+    if( it != _mapLinkIsNonColliding.end() ) {
+        return it->second;
+    }
+    return -1;
+}
+
+void Grabbed::UpdateCollidingLinks()
+{
+    KinBodyPtr pbody = RaveInterfaceCast<KinBody>(_plinkrobot->GetParent());
+    if( !pbody ) {
+        return;
+    }
+    EnvironmentBasePtr penv = pbody->GetEnv();
+    KinBodyConstPtr pgrabbedbody(_pgrabbedbody);
+    if( !pgrabbedbody || !pgrabbedbody->IsEnabled() ) {
+        _listNonCollidingLinks.clear();
+        return;
+    }
+
+    CollisionOptionsStateSaverPtr colsaver;
+    CollisionCheckerBasePtr pchecker = pbody->GetSelfCollisionChecker();
+    if( !pchecker ) {
+        pchecker = penv->GetCollisionChecker();
+    }
+
+    std::map<KinBody::LinkConstPtr, int>::iterator itnoncolliding;
+    std::vector<KinBody::LinkPtr > vbodyattachedlinks;
+    FOREACHC(itgrabbed, pbody->_vGrabbedBodies) {
+        boost::shared_ptr<Grabbed const> pgrabbed = boost::dynamic_pointer_cast<Grabbed const>(*itgrabbed);
+        bool bsamelink = find(_vattachedlinks.begin(),_vattachedlinks.end(), pgrabbed->_plinkrobot) != _vattachedlinks.end();
+        KinBodyPtr pothergrabbedbody(pgrabbed->_pgrabbedbody);
+        if( !!pothergrabbedbody && pothergrabbedbody != pgrabbedbody && pothergrabbedbody->GetLinks().size() > 0 ) {
+            if( bsamelink ) {
+                pothergrabbedbody->GetLinks().at(0)->GetRigidlyAttachedLinks(vbodyattachedlinks);
+            }
+            KinBody::KinBodyStateSaverPtr othergrabbedbodysaver;
+            FOREACHC(itgrabbedlink, pothergrabbedbody->GetLinks()) {
+                itnoncolliding = _mapLinkIsNonColliding.find(*itgrabbedlink);
+                if( itnoncolliding == _mapLinkIsNonColliding.end() ) {
+                    if( bsamelink && find(vbodyattachedlinks.begin(),vbodyattachedlinks.end(), *itgrabbedlink) != vbodyattachedlinks.end() ) {
+                        _mapLinkIsNonColliding[*itgrabbedlink] = 0;
+                    }
+                    else {
+                        // new body?
+                        if( !colsaver ) {
+                            // have to reset the collision options
+                            colsaver.reset(new CollisionOptionsStateSaver(pchecker,0));
+                        }
+                        if( !othergrabbedbodysaver ) {
+                            othergrabbedbodysaver.reset(new KinBody::KinBodyStateSaver(pothergrabbedbody, KinBody::Save_LinkEnable));
+                            pothergrabbedbody->Enable(true);
+                        }
+                        _mapLinkIsNonColliding[*itgrabbedlink] = !pchecker->CheckCollision(KinBody::LinkConstPtr(*itgrabbedlink), pgrabbedbody);
+                    }
+                }
+            }
+        }
+    }
+
+    std::set<KinBodyConstPtr> _setgrabbed;
+    FOREACHC(itgrabbed, pbody->_vGrabbedBodies) {
+        boost::shared_ptr<Grabbed const> pgrabbed = boost::dynamic_pointer_cast<Grabbed const>(*itgrabbed);
+        KinBodyConstPtr pothergrabbedbody(pgrabbed->_pgrabbedbody);
+        if( !!pothergrabbedbody ) {
+            _setgrabbed.insert(pothergrabbedbody);
+        }
+    }
+
+    _listNonCollidingLinks.clear();
+    itnoncolliding = _mapLinkIsNonColliding.begin();
+    while( itnoncolliding != _mapLinkIsNonColliding.end() ) {
+        KinBodyPtr noncollidingparent = itnoncolliding->first->GetParent(true);
+        if( !noncollidingparent ) {
+            _mapLinkIsNonColliding.erase(itnoncolliding++);
+            continue;
+        }
+        if( noncollidingparent != pbody ) {
+            // check if body is currently being grabbed
+            if( _setgrabbed.find(noncollidingparent) == _setgrabbed.end() ) {
+                _mapLinkIsNonColliding.erase(itnoncolliding++);
+                continue;
+            }
+        }
+
+        if( itnoncolliding->second && itnoncolliding->first->IsEnabled() ) {
+            _listNonCollidingLinks.push_back(itnoncolliding->first);
+        }
+        ++itnoncolliding;
     }
 }
